Config `frequency` -> `spacing`, remove vertical
