+++ Forge/src/main/java/supercoder79/survivalisland/config/ConfigData.java
@@ -1,6 +1,6 @@
 package supercoder79.survivalisland.config;
 
-import supercoder79.survivalisland.noise.OctaveNoiseRecipe;
+import supercoder79.survivalisland.noise.OctaveNoiseRecipe2D;
 import supercoder79.survivalisland.util.FloatRange;
 
 public class ConfigData {
@@ -11,8 +11,8 @@ public class ConfigData {
     public FloatRange continentalTargetRangeB = new FloatRange(-1.0f, 1.4f);
     public double islandUnderwaterFalloffDistanceMultiplier = 9;
 
-    public OctaveNoiseRecipe domainWarpNoise = new OctaveNoiseRecipe(1, 28, 24, 22, 1.732, 1.732);
-    public OctaveNoiseRecipe rangeVariationNoise = new OctaveNoiseRecipe(2, 344, 24, 1.2, 1.732, 1.732);
+    public OctaveNoiseRecipe2D domainWarpNoise = new OctaveNoiseRecipe2D(1, 28, 22, 1.732, 1.732);
+    public OctaveNoiseRecipe2D rangeVariationNoise = new OctaveNoiseRecipe2D(2, 344, 1.2, 1.732, 1.732);
 
     public boolean hardcoreMode = false;
 }
